refactor: rewrite tariancli using cobra CLI pkg
