+++ cmd/tarianctl/cmd/version.go
@@ -1,0 +1,24 @@
+package cmd
+
+import (
+	"fmt"
+
+	"github.com/spf13/cobra"
+)
+
+// nolint: gochecknoglobals
+var (
+	version    = "dev"
+	commit     = "main"
+	versionStr = version + " (" + commit + ")"
+)
+
+// versionCmd represents the version command
+var versionCmd = &cobra.Command{
+	Use:   "version",
+	Args:  cobra.NoArgs,
+	Short: "Prints version of tarianctl",
+	Run: func(cmd *cobra.Command, args []string) {
+		fmt.Printf("tarianctl version: %s\n", versionStr)
+	},
+}
