+++ cmd/tarianctl/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/kube-tarian/tarian/cmd/tarianctl/cmd"
+
+func main() {
+	cmd.Execute()
+}
