--- pkg/tarianctl/cmd/add/doc.go
@@ -1,2 +1,0 @@
-// Package add provides add commands
-package add
