--- pkg/tarianctl/cmd/get/doc.go
@@ -1,2 +1,0 @@
-// Package get provides get commands
-package get
