--- pkg/tarianctl/cmd/get/get.go
@@ -1,18 +1,0 @@
-package get
-
-import (
-	cli "github.com/urfave/cli/v2"
-)
-
-func NewGetCommand() *cli.Command {
-	return &cli.Command{
-		Name:  "get",
-		Usage: "Get resources from the Tarian Server.",
-		Flags: []cli.Flag{},
-		Subcommands: []*cli.Command{
-			NewGetConstraintsCommand(),
-			NewGetEventsCommand(),
-			NewGetActionsCommand(),
-		},
-	}
-}
