--- pkg/tarianctl/cmd/remove/doc.go
@@ -1,2 +1,0 @@
-// Package remove provides remove commands
-package remove
